Rename query files
